requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
crispy-bulma==0.8.0
dj-database-url==0.5.0
Django==4.0.6
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-markdownx==3.0.1
django-storages==1.12.3
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
Markdown==3.3.4
martor==1.6.13
Pillow==9.2.0
protobuf==4.21.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-decouple==3.6
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.10
WebOb==1.8.7
